+++ src/DependencyInjection/CompilerPass/ResourceOwnerMapCompilerPass.php
@@ -11,12 +11,13 @@
 
 namespace HWI\Bundle\OAuthBundle\DependencyInjection\CompilerPass;
 
+use HWI\Bundle\OAuthBundle\DependencyInjection\HWIOAuthExtension;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
 
 /**
- * Add resource ownermaps to the locator.
+ * Add resource ownermaps to the locator and utils.
  */
 final class ResourceOwnerMapCompilerPass implements CompilerPassInterface
 {
@@ -25,10 +26,18 @@ final class ResourceOwnerMapCompilerPass implements CompilerPassInterface
      */
     public function process(ContainerBuilder $container): void
     {
-        $def = $container->getDefinition('hwi_oauth.resource_ownermap_locator');
+        /** @var HWIOAuthExtension $extension */
+        $extension = $container->getExtension('hwi_oauth');
+        $firewallNames = $extension->getFirewallNames();
 
-        foreach ($container->getParameter('hwi_oauth.firewall_names') as $firewallId) {
-            $def->addMethodCall('set', [$firewallId, new Reference('hwi_oauth.resource_ownermap.'.$firewallId)]);
+        $locatorDef = $container->getDefinition('hwi_oauth.resource_ownermap_locator');
+        $oauthUtilsDef = $container->getDefinition('hwi_oauth.security.oauth_utils');
+
+        foreach ($firewallNames as $firewallName => $_) {
+            $resourceOwnerMapRef = new Reference('hwi_oauth.resource_ownermap.'.$firewallName);
+
+            $locatorDef->addMethodCall('set', [$firewallName, $resourceOwnerMapRef]);
+            $oauthUtilsDef->addMethodCall('addResourceOwnerMap', [$resourceOwnerMapRef]);
         }
     }
 }