+++ src/DependencyInjection/Security/Factory/OAuthAuthenticatorFactory.php
@@ -31,6 +31,13 @@ use Symfony\Component\DependencyInjection\Reference;
  */
 final class OAuthAuthenticatorFactory extends AbstractFactory implements AuthenticatorFactoryInterface, FirewallListenerFactoryInterface
 {
+    private \ArrayIterator $firewallNames;
+
+    public function __construct(\ArrayIterator $firewallNames)
+    {
+        $this->firewallNames = $firewallNames;
+    }
+
     /**
      * @param ArrayNodeDefinition $node
      */
@@ -73,6 +80,8 @@ final class OAuthAuthenticatorFactory extends AbstractFactory implements Authent
             ->addArgument(new Reference('http_kernel'))
             ->addArgument(array_intersect_key($config, $this->options))
         ;
+
+        $this->firewallNames[$firewallName] = true;
 
         return $authenticatorId;
     }
@@ -140,6 +149,8 @@ final class OAuthAuthenticatorFactory extends AbstractFactory implements Authent
             ->addArgument(new Reference('hwi_oauth.user_checker'))
             ->addArgument(new Reference('security.token_storage'))
         ;
+
+        $this->firewallNames[$id] = true;
 
         return $providerId;
     }