+++ src/DependencyInjection/Security/Factory/OAuthFactory.php
@@ -29,6 +29,13 @@ use Symfony\Component\DependencyInjection\Reference;
  */
 final class OAuthFactory extends AbstractFactory
 {
+    private \ArrayIterator $firewallNames;
+
+    public function __construct(\ArrayIterator $firewallNames)
+    {
+        $this->firewallNames = $firewallNames;
+    }
+
     /**
      * @param ArrayNodeDefinition $node
      */
@@ -99,6 +106,8 @@ final class OAuthFactory extends AbstractFactory
         ;
 
         $this->createRefreshTokenListeners($container, $providerId, $id);
+
+        $this->firewallNames[$id] = true;
 
         return $providerId;
     }